+++ springbootAPI/src/main/java/net/javaguides/springboot/controller/SupplierController.java
@@ -34,21 +34,21 @@ public class SupplierController {
     // build get supplier by id REST API
     // http://localhost:8080/api/suppliers/1
     @GetMapping("{id}")
-    public ResponseEntity<Supplier> getSupplierById(@PathVariable("id") long supplierId){
+    public ResponseEntity<Supplier> getSupplierById(@PathVariable("id") int supplierId){
         return new ResponseEntity<Supplier>(supplierService.getSupplierById(supplierId), HttpStatus.OK);
     }
 
     // build update supplier REST API
     // http://localhost:8080/api/suppliers/1
     @PutMapping("{id}")
-    public ResponseEntity<Supplier> updateSupplier(@PathVariable("id") long id, @RequestBody Supplier supplier){
+    public ResponseEntity<Supplier> updateSupplier(@PathVariable("id") int id, @RequestBody Supplier supplier){
         return new ResponseEntity<Supplier>(supplierService.updateSupplier(supplier, id), HttpStatus.OK);
     }
 
     // build delete supplier REST API
     // http://localhost:8080/api/suppliers/1
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteSupplier(@PathVariable("id") long id){
+    public ResponseEntity<String> deleteSupplier(@PathVariable("id") int id){
 
         // delete supplier from DB
         supplierService.deleteSupplier(id);
