+++ springbootAPI/src/main/java/net/javaguides/springboot/repository/SupplierRepository.java
@@ -3,6 +3,6 @@ package net.javaguides.springboot.repository;
 import net.javaguides.springboot.model.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupplierRepository extends JpaRepository<Supplier, Long> {
+public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
 }
