+++ springbootAPI/src/main/java/net/javaguides/springboot/service/CategoryService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface CategoryService {
     Category saveCategory(Category category);
     List<Category> getAllCategories();
-    Category getCategoryById(long id);
-    Category updateCategory(Category category, long id);
-    void deleteCategory(long id);
+    Category getCategoryById(int id);
+    Category updateCategory(Category category, int id);
+    void deleteCategory(int id);
 }
